tg_todo/requirements.txt
asgiref==3.8.0
attrs==23.2.0
Django==5.0.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.4.4
uritemplate==4.1.1
